+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @autotranslate/vite
 
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @autotranslate/cli@1.0.0-beta.5
+
 ## 1.0.0-beta.4
 
 ### Patch Changes